# docs/openapi.yaml
openapi: 3.1.0
info:
  title: COGNOS Accounts & Multi-Tenant Workspaces (Phase 24)
  version: "24.0"
  summary: Email + Google sign-in, isolated append-only workspaces over the trust-annotated Atlas.
  description: |
    Multi-tenant workspace layer over the Phase 23 trust-annotated knowledge
    graph ("Atlas").

    ## Provenance rule
    User facts are asserted ONLY from trusted Atlas rows:

      * `[graph_mu02bqyvnd1cmlmv]` — user.name.full "Jeremy Bryan Perritt" (trusted, truth-bearing)
      * `[graph_mu02bqun6vptjczp]` — user.role, developer of the assistant (trusted)
      * `[graph_mu02br2ofvew5mqm]` — user.preferred.name "Patches" (trusted; seeds the default display_name)
      * `[graph_mu02br6ged2v5219]` — user.values: autonomy, deep work, maker mindset, subtractive innovation, opposes bureaucracy & merit-ocracy (trusted)
      * `[graph_mu02bra9qiw2n5kp]` — user.name.variant "Jeremy Brian Perritt" (trusted)
      * `[graph_mu01yzi4hosh3erz]` — user.name.middle spelled "Bryan" (trusted)
      * `[graph_mtzt7k3ei9ozy2ci]` — conservative duplicate of the name (trusted)
      * `[graph_mtzt7jzlg5oklort]` — assistant no longer tied to the Base 44 integration layer (trusted)

    Rows marked NOT truth-bearing are contextual only and never generate
    assertions.

    ## Auth
    Bearer JWT (HS256): claims `sub` (user_id), `ws` (workspace_id — the only
    workspace the bearer may touch), `jti` (revocable via logout blocklist).
...

    ## Sealed provenance
...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing, invalid, or revoked token (401).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
...
    Bucket:
      type: object
      properties:
        namespace:
          type: string
          example: group-public-3f1d6c2e-8f6a-4c1d-9d3a-2f4b5c6d7e8f
        read:
          type: boolean
        write:
          type: boolean
    Group:
      type: object
      properties:
        group_id:
          type: string
          format: uuid
        group_name:
          type: string
        created_at:
          type: string
          format: date-time
        your_role:
          type: string
          enum: [owner, member]
        buckets:
          type: object
          properties:
            public:
              $ref: '#/components/schemas/Bucket'
            private:
              $ref: '#/components/schemas/Bucket'
        owner:
          type: object
          properties:
            id:
              type: string
            email:
              type: [string, "null"]
              format: email
        members:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              email:
                type: [string, "null"]
                format: email
              can_write_private:
                type: boolean
paths:
...
  /api/workspaces:
    get:
      tags: [workspaces]
      summary: Everything the bearer can see — private namespace plus group buckets with the role matrix applied.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/groups:
    get:
      tags: [groups]
      summary: Groups the bearer belongs to (with bucket permissions).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK.
    post:
      tags: [groups]
      summary: Create a group; the creator is the owner; both buckets materialize immediately.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 120
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
...
security: []